+++ cxroots/root_finding_demo.py
@@ -6,7 +6,6 @@ from .root_finding import find_roots_gen
 def _update_frame(frame, original_contour):
     import matplotlib.pyplot as plt
 
-    fig = plt.gcf()
     ax = plt.gca()
 
     roots, _, boxes, num_remaining_roots = frame
@@ -23,13 +22,46 @@ def _update_frame(frame, original_contour):
 
     plt.scatter(np.real(roots), np.imag(roots), color="k", marker="x")
     ax.text(
-        0.02,
-        0.95,
-        "Zeros remaining: %i" % num_remaining_roots,
-        transform=ax.transAxes,
+        0.02, 0.95, "Zeros remaining: %i" % num_remaining_roots, transform=ax.transAxes
     )
     original_contour._size_plot()
-    fig.canvas.draw()
+    plt.draw()
+
+
+def demo_roots_animation(original_contour, f, df=None, **roots_kwargs):
+    """
+    Creates an animation object where each frame is a step in the rootfinding process
+
+    Parameters
+    ----------
+    original_contour : :class:`Contour <cxroots.contour.Contour>`
+        The contour which bounds the region in which all the roots of
+        f(z) are sought.
+    f : function
+        A function of a single complex variable, z, which is analytic
+        within the contour and has no poles or roots on the contour.
+    df : function, optional
+        A function of a single complex variable which is the derivative
+        of the function f(z). If df is not given then it will be
+        approximated with a finite difference formula.
+    **roots_kwargs : kwargs
+        Other keyword arguments to pass to
+        :func:`find_roots_gen <cxroots.root_finding.find_roots_gen>`
+
+    Returns
+    -------
+    animation.FuncAnimation
+        An animation where each frame is a step in the rootfinding process
+    """
+    import matplotlib.pyplot as plt
+    from matplotlib import animation
+
+    fig = plt.gcf()
+
+    root_finder = find_roots_gen(original_contour, f, df, **roots_kwargs)
+    return animation.FuncAnimation(
+        fig, _update_frame, frames=root_finder, fargs=[original_contour]
+    )
 
 
 def demo_find_roots(
@@ -38,7 +70,6 @@ def demo_find_roots(
     df=None,
     save_file=None,
     auto_animation=False,
-    return_animation=False,
     writer=None,
     **roots_kwargs
 ):
@@ -54,41 +85,34 @@ def demo_find_roots(
         If False (default) then press SPACE to step the animation forward
         If True then the animation will play automatically until all the
         roots have been found.
-    return_animation : bool, optional
-        If True then the matplotlib animation object will be returned
-        instead of being shown.  Defaults to False.
     writer : str, optional
         Passed to :meth:`matplotlib.animation.FuncAnimation.save`.
     **roots_kwargs
         Additional key word arguments passed to :meth:`~cxroots.contour.Contour.roots`.
     """
     import matplotlib.pyplot as plt
-    from matplotlib import animation
 
-    fig = plt.gcf()
+    if save_file or auto_animation:
+        anim = demo_roots_animation(original_contour, f, df=None, **roots_kwargs)
 
-    root_finder = find_roots_gen(original_contour, f, df, **roots_kwargs)
-
-    if save_file or auto_animation or return_animation:
-        anim = animation.FuncAnimation(
-            fig, _update_frame, frames=root_finder, fargs=[original_contour]
-        )
-
-    if return_animation:
-        return anim
-    elif save_file:
+    if save_file:
         anim.save(filename=save_file, fps=1, dpi=200, writer=writer)
         plt.close()
     elif auto_animation:
         plt.show()
     else:
         # Create event to handler to let user move through frames
+        root_finder = find_roots_gen(original_contour, f, df, **roots_kwargs)
+        original_contour.plot(linecolor="k", linestyle="--")
 
         def draw_next(event):
             if event.key == " ":
-                _update_frame(next(root_finder))
+                try:
+                    _update_frame(next(root_finder), original_contour)
+                except StopIteration:
+                    # No more roots to find
+                    pass
 
+        fig = plt.gcf()
         fig.canvas.mpl_connect("key_press_event", draw_next)
         plt.show()
-
-    return None